+++ CHANGELOG.md
@@ -104,6 +104,7 @@ const instance = new Instance({
 
 ### Fix
 
+- **GlobeControls**: pass the ellipsoid to the underlying controls (#657)
 - **TileFS.glsl**: fix non working shadow maps on MacOS (#579)
 - **PointCloud**: fix filters not being applied
 - **Tiles3D**: correctly handle load-error from 3DtilesRenderer